First iteration of implementing a dashboard.
We introduced an aggregate which has data and behaviour.

All behaviour emits internal events. Which then mutate the data.

This, however, is too much for this stage. We need to rewind a bit:
We dont't want aggregates, internal events and so on, yet.

We just want events and a model that can handle those. So let's KISS and
remove most in a next commit