Introduce a simple dashboard feature.
We create events. And then pass those along into a Dashboard.

This Dashboard can handle those events and build a dashboard from it.
For now, it merely adds the "total buying price". Nothing fancy yet.

The handler is a tad more complex than needed, but that is to show how
we will handle differnt types of events in future. And it allows us to
focus a little better on the business-logic and less on how to loop over
events and do stuff with them.