Add a simple dividend feature.
We can see now that events events in the event stream follow eachother,
and can therefore depend on eachother. The events "StocksBought"
increase the amount of stocks for a certain kind, one has.
The event DividendPaid then uses that state to check "how many stocks of
X does the portfolio hold" to calculate the total dividend paid.

This lacks domain-feature of "ex dividend" date, so it is technically
not correct yet
Many brokers and overviews will report only the total "dividend paid"
amount, and not the amount-per-stock. Our approach assumes the amounts
are "per stock". This is a choice we can make, as developers.